split long file and create linera-execution/src/resources.rs
